requirements.txt
openai-agents
fastmcp
duckduckgo-search
python-dotenv
